search_engine/requirements.txt
# ===========================================
# SelmaData Project - Production-Grade Dependencies
# ===========================================

# Core LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-experimental>=0.3.0
langchain-huggingface>=0.1.0

# Document Loading & Processing
pypdf>=4.0.0
pdfplumber>=0.11.0
python-docx>=1.1.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0
chardet>=5.2.0
unstructured>=0.12.0

# Embeddings - Local & Free
sentence-transformers>=2.2.0
torch>=2.0.0

# Vector Store - Local
chromadb>=0.5.0

# Re-ranking - Local & Free (pin version for API stability)
flashrank==0.2.9

# HaciProvider CoreProcessor API
haci_provider>=2.0.0

# SelmaData Evaluation (pin to stable 0.1.x for API compatibility)
selma_dataas==0.1.21
datasets>=2.14.0

# LangChain Provider (for SelmaDataAs evaluation)
langchain-openai>=0.2.0

# Configuration Management
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Web Interface
streamlit>=1.30.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
tqdm>=4.66.0
rich>=13.0.0
loguru>=0.7.0
tenacity>=8.2.0
numpy>=1.24.0
pandas>=2.0.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Type Checking
types-requests>=2.31.0
types-beautifulsoup4>=4.12.0
